Ensure that the function declaration starts at the beginning of the line
Since forever we just did some regexp matching of the function
expression with return type and parameters.

But we need to take into account the beginning of the line as well as
code which contains a valid function declaration like

SomeStuff Function notAFunc()

is not one.

We now search for the only valid modifiers before it as well. We don't
enforce a valid subset of these though.